% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{kostenstring_zu_numerisch}
\alias{kostenstring_zu_numerisch}
\title{Deutsche Kommazahl als String zu numerisch}
\usage{
kostenstring_zu_numerisch(x)
}
\description{
Einen String in deutschem zahklenformat, z.B. 123,12 in 
numeric umwandeln. Wird übersprungen falls wert schon nuermisch ist,
}
\keyword{internal}
